[CL-2115] Move config
