+++ front/app/components/SignUpIn/SignUp/ConfirmationSignupStep/index.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState, FormEvent } from 'react';
+import React, { useState, FormEvent } from 'react';
 import { SignUpStepOutletProps } from 'utils/moduleUtils';
 import { FormattedMessage } from 'utils/cl-intl';
 import { trackEventByName } from 'utils/analytics';
@@ -6,7 +6,7 @@ import tracks from './tracks';
 import messages from './messages';
 import Error from 'components/UI/Error';
 import { confirm, resendCode, IConfirmation } from 'services/confirmation';
-import useAuthUser, { TAuthUser } from 'hooks/useAuthUser';
+import useAuthUser from 'hooks/useAuthUser';
 import { isNilOrError } from 'utils/helperUtils';
 import { CLErrors, CLError } from 'typings';
 import styled from 'styled-components';
@@ -23,8 +23,6 @@ import {
 import Link from 'utils/cl-router/Link';
 import Button from 'components/UI/Button';
 import { FormLabel } from 'components/UI/FormComponents';
-
-const CONFIRMATION_STEP_NAME = 'confirmation';
 
 const FormContainer = styled.div<{ inModal: boolean }>`
   display: flex;
@@ -100,13 +98,9 @@ const FooterNoteSuccessMessageIcon = styled(Icon)`
   margin-right: 4px;
 `;
 
-type Props = Pick<SignUpStepOutletProps, 'onCompleted' | 'onData' | 'step'>;
+type Props = Pick<SignUpStepOutletProps, 'onCompleted' | 'step'>;
 
-const userEmailToBeConfirmed = (authUser: TAuthUser) => {
-  return !isNilOrError(authUser) && authUser.attributes.confirmation_required;
-};
-
-const ConfirmationSignupStep = ({ onCompleted, onData, step }: Props) => {
+const ConfirmationSignupStep = ({ onCompleted, step }: Props) => {
   const user = useAuthUser();
   const [confirmation, setConfirmation] = useState<IConfirmation>({
     code: null,
@@ -117,21 +111,7 @@ const ConfirmationSignupStep = ({ onCompleted, onData, step }: Props) => {
   const [changingEmail, setChangingEmail] = useState(false);
   const [codeResent, setCodeResent] = useState(false);
 
-  useEffect(() => {
-    onData({
-      key: CONFIRMATION_STEP_NAME,
-      position: 4,
-      stepDescriptionMessage: messages.confirmYourAccount,
-      isEnabled: (authUser) => {
-        return userEmailToBeConfirmed(authUser);
-      },
-      isActive: userEmailToBeConfirmed,
-      canTriggerRegistration: true,
-    });
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  if (step !== CONFIRMATION_STEP_NAME || isNilOrError(user)) {
+  if (isNilOrError(user)) {
     return null;
   }
